ARM64: Update types for (in)equality checks
For equalities and inequalities, we can change the types on each side to
always convert to the larger type, which ensures all conversions are
safe.

I've done this in a separate commit because my dev setup doesn't use
ARM. In case something ARM-specific breaks in CI, I'd like to be able to
keep those changes in their own commit so I can debug more easily.

+++ src/vmm/src/arch/aarch64/regs.rs
@@ -843,7 +843,7 @@ mod tests {
 
         let reg_ref: Aarch64RegisterRef = (&old_reg).try_into().unwrap();
         assert_eq!(old_reg.id, reg_ref.id);
-        assert_eq!(old_reg.data as u64, reg_ref.value::<u64, 8>());
+        assert_eq!(old_reg.data, u128::from(reg_ref.value::<u64, 8>()));
 
         let old_reg = Aarch64RegisterOld {
             id: KVM_REG_SIZE_U256,
@@ -870,7 +870,7 @@ mod tests {
 
         let reg: Aarch64RegisterOld = reg_ref.try_into().unwrap();
         assert_eq!(reg.id, reg_ref.id);
-        assert_eq!(reg.data as u64, reg_ref.value::<u64, 8>());
+        assert_eq!(reg.data, u128::from(reg_ref.value::<u64, 8>()));
 
         let reg_ref = Aarch64RegisterRef::new(KVM_REG_SIZE_U256, &[0_u8; 32]);
